feat: add ut、contractInfo manager
